module11/notification_service/requirements.txt
# Core framework
fastapi==0.115.2
uvicorn[standard]==0.32.0

# Security and JWT
python-jose==3.3.0
passlib[bcrypt]==1.7.4  

# Redis (async)
redis==5.0.3

# Async utilities
aiofiles==23.2.1

# WebSocket and typing support
starlette==0.40.0  # required explicitly for WebSocketState import
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator
pydantic[email]
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-sqlalchemy 
opentelemetry-instrumentation-redis 
opentelemetry-exporter-otlp-proto-grpc